requirements.txt
streamlit==1.22.0
PyPDF2==3.0.1
requests==2.31.0
Pillow==9.5.0
python-dotenv==1.0.0